Complete Queues and fix mistake with peek in Stack
Complete the first video on Queues, one with an array based implementation and a
second implementation with two stacks. I will be completing the challenge for queues
later. Also fixed a mistake with my Stack implementation of peek, which was looking
at the first element entered into the list instead of the last element inserted.